BIG FIX. persistence issue: phantom responses count. even when you add a new book, it shows that some ideas already have responses, which is false. fix is to attach the idea id to a unique book id. we go from i1, i2, i3,... only to b1i1, b1i2, b1i3...
Root cause: The problem is that idea IDs are generated as i1, i2, i3, etc. in a sequential pattern. When a new book is created, it gets ideas with the same IDs (i1, i2, i3) as previous books. So when the app queries for responses with ideaId == i1, it finds responses from ALL previous books that had an idea with ID i1.